Refactor: Implement some of the learnings from Astra
Initially I attempted to use the package with Astra and it's command
line formatting needs but it fell short in a number of places.

This refactor adds a new Tree and Spinner implementation (instead of
just "pretty") it also makes it so users can include filter directives
as part of macros.