fix: Spacing and formatting issues within line length formatting
Previously the logic to do proper line length formatting did so in a fashion that separated the text
by whitespace and used this whitespace split to figure out where the formatter should make the cut on
text.

This worked okay until the user wanted to do something a little creative with the formatting of the
text, namely splitting the text into multiple lines and indenting some of those lines.

This fix introduces the ability to preserve some of that formatting, such that the user can format
to their liking and still have word wrap support.