feat: Add empty case for macros
Previously you could not leave the formatting macros empty due to there not being a macro case for it. This is useful when calling things like println! and expecting a new line to be printed without having to put in any text.

This also adds more test cases that will eventually be fashioned into a golden file test.