fix: Only method is logically incorrect
There was a bug where the only method check for "allowed methods" in
the JSON module was instead inversiing itself. This made it so that
using only while using the JSON type would print the in json even
if you told it to only print another format and then if you told it
to print json it would stop printing for every other instance of print.