fix: properly handle double newline wrapping
Previously the auto-wrapping feature would treat double newlines as
a single word and therefore just inject it into the text. This worked
most of the time but it was not correct and broke some thing like the
tree module.

This changes the text parsing to instead treat double newlines as
separate and act accordingly. This doesnt' immediatley fix tree, but
its the first step to improving it.