feat: Add configurable output targets and builder options
Previously we wrote directly to stdout no matter what.

This updates the Options struct such that the user can change which
output target. This allows programs to be more flexible and output
to not just stdout but maybe to a file or a bufwriter. This also
helps with testing since we can just output from the target instead
of trying to capture stdout.

Also added builder style methods for each of Option's fields.